requirements.txt
asgiref              3.6.0
Django               4.1.7
django-adminlte-3    0.1.6
django-restframework 0.0.1
djangorestframework  3.14.0
Pillow               9.4.0
pip                  23.0.1
pytz                 2023.2
setuptools           67.4.0
sqlparse             0.4.3
tzdata               2023.2
wheel                0.38.4